Request.Error

Description

The Error Module, containing predefined errors. These can occur, when a Request is parsed.

Documentation

data ParseError

Constructors for different kinds of error.

Constructors

RequestLineMalformed ByteString
Error, when the first line of the request is not correctly formated.

HttpMethodNotSupported ByteString
Error, when the Http-Method of the request is not supported.

HttpVersionNotSupported ByteString
Error, when the Http Version of the request is not supported.

UnknownParseError ByteString