Request.Internal.Request

Description

bundles internal functions and datatypes to parse a Http-Request.

Documentation

parseRequest :: Handle -> IO Request

top level function, that is used to parse a request from a handle.

parseRequestFromString :: [ByteString] -> Request

Parses a request from a bytestring.

parseToString :: Handle -> [ByteString] -> IO [ByteString]

parses the request from a handle line by line to a list of bytesstrings.

parseHeaders :: [ByteString] -> [(ByteString, [ByteString])]

parses the headers from a list of bytestrings, that contain the lines.

parseHeader :: ByteString -> (ByteString, [ByteString])

reads one header to a tupel (headername, listOfValues)

fromString :: ByteString -> ParseMonad RequestLine

Top level function for parsing a byteString to a RequestLine.

data Request

The constructor of the request. A request contains a requestline (methond, path and http-version) and a list of headers.

Constructors

Request

Fields

requestLine :: RequestLine

headers :: [(ByteString, [ByteString])]

data RequestLine

Constructors

RequestLine

Fields

method :: ByteString

path :: ByteString

version :: ByteString