Response.Response

Description

Module modeling an HTTP Response and its transformation to a BysteString.

Documentation

data Response

Constructors

Response

Fields

statusCode :: StatusCode

version :: ByteString

headers :: Headers

content :: ByteString

getStatusCode :: Response -> Integer

Extract the response status code.

getReasonPhrase :: Response -> String

Extract the response reaso phrase.

getVersion :: Response -> ByteString

Extract the http version.

getHeaders :: Response -> Headers

Extract the response headers.

getContent :: Response -> ByteString

Extract the response content

buildOkResponse :: Headers -> ByteString -> Response

Template for building an OK response where only the request content is missing.

buildNotFoundResponse :: Headers -> ByteString -> Response

Template for building a NotFound response where only the request content is missing.

buildInternalServerErrorResponse :: Headers -> ByteString -> Response

Template for building a InternalServerError Response where only the request content is missing.

toByteString :: Response -> ByteString

Transform a response to its ByteString representation.